Add Special* groups to show known identifiers in bold
This commit adds SpecialConstant, SpecialIdentifier, and SpecialMacro
to make known constants, identifiers, and macros in bold.

In everyday programming, we works upon frameworks
and project-specific libraries, which provide
global variables, constants, and macros,
something like "expansion packs."

Keywords in these expansion packages are
already highlighted by general rules.
But it will be much better if we can figure out
they are from expansion packs, because
we can tell what is local, what is from framework,
and what might be a typo.

This commit adds Special-series groups,
they are the same as their original version,
but in bold.  They are in the same colors,
so the user experience is the same.
But they are in bold, so you can know it's from a framework.

Signed-off-by: Sung-Yu Chen <sychen@realtek.com>